requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.1
beautifulsoup4>=4.12.2
readability-lxml>=0.8.1
lxml>=4.9.3
redis>=5.0.0
slowapi>=0.1.9
pytest>=7.4.0
openai>=1.0.0
python-dotenv>=1.0.0
